Piecewise override some segment boundary checking
It appears that some models (14 out of 141 cases) may have incorrectly emulated
EEPROM sizes. In ten cases, writes past the end of the EEPROM advertised size
are preserved, and in four other ones, writes within the advertised boundary
are not preserved. For now, set up a manual set of overrides to prevent the
past-the-end checks for the specific cases known to fail, so that the other 90%
of the checks can continue to be run.