Improve the mechanism for skipping tests
The general intent has been, and continues to be, to test whatever can be
successfully tested, without failing on things that may not be under our
control.

Previously, overriding of Make rules enabled this behavior, but spewed ugly
warnings from Make, and there was no effective regular structure to the
overrides. Now, if an entire test is to be skipped, an environment variable can
be set, whether in an overrides file or otherwise, which increases flexibility
and provides a more regular structure to the overrides files.

There is, however, no requirement that overrides files use only this means of
skipping tests. Indeed, some overrides files may not skip tests at all.